remove layout stuff
