+++ src/app/Components/ArtworkRail/ArtworkRail.tsx
@@ -3,7 +3,6 @@ import {
   ArtworkRail_artworks$data,
   ArtworkRail_artworks$key,
 } from "__generated__/ArtworkRail_artworks.graphql"
-import { SellWithArtsyRecentlySold_recentlySoldArtworkTypeConnection$data } from "__generated__/SellWithArtsyRecentlySold_recentlySoldArtworkTypeConnection.graphql"
 import {
   ARTWORK_RAIL_CARD_IMAGE_HEIGHT,
   ArtworkRailCard,
@@ -28,26 +27,23 @@ import { graphql, useFragment } from "react-relay"
 const MAX_NUMBER_OF_ARTWORKS = 30
 export const ARTWORK_RAIL_IMAGE_WIDTH = 295
 
-interface CommonArtworkRailProps {
+export interface ArtworkRailProps extends ArtworkActionTrackingProps {
+  artworks: ArtworkRail_artworks$key
+  onPress?: (artwork: ArtworkRail_artworks$data[0], index: number) => void
   dark?: boolean
   hideArtistName?: boolean
+  hideCuratorsPickSignal?: boolean
+  hideIncreasedInterestSignal?: boolean
   showPartnerName?: boolean
   ListFooterComponent?: ReactElement | null
   ListHeaderComponent?: ReactElement | null
   listRef?: React.RefObject<FlatList<any>>
   onEndReached?: () => void
   onEndReachedThreshold?: number
-  showSaveIcon?: boolean
   onMorePress?: () => void
-  viewabilityConfig?: ViewabilityConfig | undefined
   onViewableItemsChanged?: (info: { viewableItems: any[]; changed: any[] }) => void
-  hideIncreasedInterestSignal?: boolean
-  hideCuratorsPickSignal?: boolean
-}
-
-export interface ArtworkRailProps extends CommonArtworkRailProps, ArtworkActionTrackingProps {
-  artworks: ArtworkRail_artworks$key
-  onPress?: (artwork: ArtworkRail_artworks$data[0], index: number) => void
+  showSaveIcon?: boolean
+  viewabilityConfig?: ViewabilityConfig | undefined
 }
 
 export const ArtworkRail: React.FC<ArtworkRailProps> = ({
@@ -55,8 +51,8 @@ export const ArtworkRail: React.FC<ArtworkRailProps> = ({
   onPress,
   onEndReached,
   onEndReachedThreshold,
-  ListHeaderComponent = SpacerComponent,
-  ListFooterComponent = SpacerComponent,
+  ListHeaderComponent = <Spacer x={2} />,
+  ListFooterComponent = <Spacer x={2} />,
   hideArtistName = false,
   showPartnerName = true,
   dark = false,
@@ -137,67 +133,6 @@ export const ArtworkRail: React.FC<ArtworkRailProps> = ({
   )
 }
 
-type RecentlySoldArtwork = NonNullable<
-  NonNullable<SellWithArtsyRecentlySold_recentlySoldArtworkTypeConnection$data["edges"]>[0]
->["node"]
-
-export interface RecentlySoldArtworksRailProps extends CommonArtworkRailProps {
-  recentlySoldArtworks: RecentlySoldArtwork[]
-  onPress?: (recentlySoldArtwork: RecentlySoldArtwork, index: number) => void
-}
-
-export const RecentlySoldArtworksRail: React.FC<RecentlySoldArtworksRailProps> = ({
-  listRef,
-  onPress,
-  onEndReached,
-  onEndReachedThreshold,
-  ListHeaderComponent = SpacerComponent,
-  ListFooterComponent = SpacerComponent,
-  hideArtistName = false,
-  recentlySoldArtworks,
-  showPartnerName = true,
-}) => {
-  return (
-    <PrefetchFlatList
-      onEndReached={onEndReached}
-      onEndReachedThreshold={onEndReachedThreshold}
-      prefetchUrlExtractor={(item) => item?.artwork?.href || undefined}
-      listRef={listRef}
-      horizontal
-      ListHeaderComponent={ListHeaderComponent}
-      ListFooterComponent={ListFooterComponent}
-      ItemSeparatorComponent={() => <Spacer x="15px" />}
-      showsHorizontalScrollIndicator={false}
-      // We need to set the maximum number of artworks to not cause layout shifts
-      data={recentlySoldArtworks.slice(0, MAX_NUMBER_OF_ARTWORKS)}
-      initialNumToRender={MAX_NUMBER_OF_ARTWORKS}
-      contentContainerStyle={{ alignItems: "flex-end" }}
-      renderItem={({ item, index }) => {
-        if (!item?.artwork) {
-          return null
-        }
-
-        return (
-          <ArtworkRailCard
-            artwork={item.artwork}
-            onPress={() => {
-              onPress?.(item, index)
-            }}
-            priceRealizedDisplay={item?.priceRealized?.display || ""}
-            lowEstimateDisplay={item?.lowEstimate?.display || ""}
-            highEstimateDisplay={item?.highEstimate?.display || ""}
-            performanceDisplay={item?.performance?.mid ?? undefined}
-            showPartnerName={showPartnerName}
-            isRecentlySoldArtwork
-            hideArtistName={hideArtistName}
-          />
-        )
-      }}
-      keyExtractor={(item, index) => String(item?.artwork?.slug || index)}
-    />
-  )
-}
-
 const artworksFragment = graphql`
   fragment ArtworkRail_artworks on Artwork @relay(plural: true) {
     ...ArtworkRailCard_artwork @arguments(width: 590)
@@ -213,8 +148,6 @@ const artworksFragment = graphql`
     }
   }
 `
-
-const SpacerComponent = () => <Spacer x={2} />
 
 export const ArtworkRailPlaceholder: React.FC = () => (
   <Join separator={<Spacer x="15px" />}>
