+++ src/app/Components/ContextMenu/ContextMenuArtworkPreviewCard.tsx
@@ -37,7 +37,7 @@ export const ContextMenuArtworkPreviewCard: React.FC<ContextMenuArtworkPreviewCa
     dark = false,
     hideArtistName = false,
     showPartnerName = true,
-    isRecentlySoldArtwork = false,
+    displayRealizedPrice = false,
     lotLabel,
     lowEstimateDisplay,
     highEstimateDisplay,
@@ -63,11 +63,11 @@ export const ContextMenuArtworkPreviewCard: React.FC<ContextMenuArtworkPreviewCa
   const backgroundColor = dark ? "black100" : "white100"
 
   const getTextHeight = () => {
-    return ARTWORK_RAIL_TEXT_CONTAINER_HEIGHT + (isRecentlySoldArtwork ? 50 : 0)
+    return ARTWORK_RAIL_TEXT_CONTAINER_HEIGHT + (displayRealizedPrice ? 50 : 0)
   }
 
   const containerWidth = FULL_WIDTH_RAIL_CARD_IMAGE_WIDTH
-  const displayForRecentlySoldArtwork = !!isRecentlySoldArtwork
+  const displayForRecentlySoldArtwork = !!displayRealizedPrice
 
   return (
     <Flex backgroundColor={backgroundColor} m={1}>
@@ -137,7 +137,7 @@ export const ContextMenuArtworkPreviewCard: React.FC<ContextMenuArtworkPreviewCa
               {partner?.name}
             </Text>
           )}
-          {!!isRecentlySoldArtwork && (
+          {!!displayRealizedPrice && (
             <RecentlySoldCardSection
               priceRealizedDisplay={priceRealizedDisplay}
               lowEstimateDisplay={lowEstimateDisplay}
@@ -147,7 +147,7 @@ export const ContextMenuArtworkPreviewCard: React.FC<ContextMenuArtworkPreviewCa
             />
           )}
 
-          {!!saleMessage && !isRecentlySoldArtwork && (
+          {!!saleMessage && !displayRealizedPrice && (
             <Text
               lineHeight="20px"
               variant="xs"
@@ -168,7 +168,7 @@ export interface ContextMenuArtworkPreviewCardImageProps {
   image: ArtworkRailCard_artwork$data["image"]
   urgencyTag?: string | null
   containerWidth?: number
-  isRecentlySoldArtwork?: boolean
+  displayRealizedPrice?: boolean
   /** imageHeightExtra is an optional padding value you might want to add to image height
    * When using large width like with RecentlySold, image appears cropped
    * TODO: - Investigate why
@@ -178,7 +178,7 @@ export interface ContextMenuArtworkPreviewCardImageProps {
 
 export const ContextMenuArtworkPreviewCardImage: React.FC<
   ContextMenuArtworkPreviewCardImageProps
-> = ({ image, urgencyTag = null, containerWidth, isRecentlySoldArtwork, imageHeightExtra = 0 }) => {
+> = ({ image, urgencyTag = null, containerWidth, displayRealizedPrice, imageHeightExtra = 0 }) => {
   const color = useColor()
   const FULL_WIDTH_RAIL_CARD_IMAGE_WIDTH = useFullWidth()
 
@@ -201,7 +201,7 @@ export const ContextMenuArtworkPreviewCardImage: React.FC<
       height: height ?? 0,
     },
     {
-      width: isRecentlySoldArtwork
+      width: displayRealizedPrice
         ? FULL_WIDTH_RAIL_CARD_IMAGE_WIDTH
         : ARTWORK_LARGE_RAIL_CARD_IMAGE_WIDTH,
       height: ARTWORK_RAIL_CARD_IMAGE_HEIGHT,
